# vitis-vision-front-end/src/openapi/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Vitis Vision Service",
    "description": "OpenAPI documentation for Vitis Vision Service",
    "contact": {
      "name": "Vitis Vision",
      "url": "https://vitisvision.com",
      "email": "vitisvision@gmail.com"
    },
    "version": "0.0.1"
  },
  "servers": [
    {
      "url": "http://localhost:8181",
      "description": "Local Server"
    }
  ],
  "security": [
    {
      "Bearer Auth": []
    }
  ],
...
  "paths": {
...
    "/api/v1/vines/{vineId}/devices/{deviceId}": {
      "put": {
        "tags": [
          "Device"
        ],
        "summary": "Update the device details",
        "description": "Update the device details in a vineyard with the provided details",
        "operationId": "updateDevice",
        "parameters": [
          {
            "name": "vineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDeviceResponse"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Device"
        ],
        "summary": "Delete the device by id",
        "description": "Delete the device by id in a vineyard with the provided details",
        "operationId": "deleteDevice",
        "parameters": [
          {
            "name": "vineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/vines/{vineId}/devices": {
      "get": {
        "tags": [
          "Device"
        ],
        "summary": "Get the device details by id",
        "description": "Get the device details by id in a vineyard with the provided details",
        "operationId": "getDevice",
        "parameters": [
          {
            "name": "vineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDeviceResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Device"
        ],
        "summary": "Create a new device in a vine",
        "description": "Create a new device in a vine with the provided details",
        "operationId": "createDevice",
        "parameters": [
          {
            "name": "vineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DeviceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDeviceResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vines/{vineId}/devices/{deviceId}/deactivate": {
      "post": {
        "tags": [
          "Device"
        ],
        "summary": "Deactivate the device",
        "description": "Deactivate the device in a vineyard with the provided details",
        "operationId": "deactivateDevice",
        "parameters": [
          {
            "name": "vineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/vines/{vineId}/devices/{deviceId}/activate/{frequency}": {
      "post": {
        "tags": [
          "Device"
        ],
        "summary": "Activate the device with the provided frequency",
        "description": "Activate the device with the provided frequency in a vineyard with the provided details",
        "operationId": "activateDevice",
        "parameters": [
          {
            "name": "vineId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "deviceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          },
          {
            "name": "frequency",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDeviceActivateResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "100 CONTINUE",
              "101 SWITCHING_PROTOCOLS",
              "102 PROCESSING",
              "103 EARLY_HINTS",
              "103 CHECKPOINT",
              "200 OK",
              "201 CREATED",
              "202 ACCEPTED",
              "203 NON_AUTHORITATIVE_INFORMATION",
              "204 NO_CONTENT",
              "205 RESET_CONTENT",
              "206 PARTIAL_CONTENT",
              "207 MULTI_STATUS",
              "208 ALREADY_REPORTED",
              "226 IM_USED",
              "300 MULTIPLE_CHOICES",
              "301 MOVED_PERMANENTLY",
              "302 FOUND",
              "302 MOVED_TEMPORARILY",
              "303 SEE_OTHER",
              "304 NOT_MODIFIED",
              "305 USE_PROXY",
              "307 TEMPORARY_REDIRECT",
              "308 PERMANENT_REDIRECT",
              "400 BAD_REQUEST",
              "401 UNAUTHORIZED",
              "402 PAYMENT_REQUIRED",
              "403 FORBIDDEN",
              "404 NOT_FOUND",
              "405 METHOD_NOT_ALLOWED",
              "406 NOT_ACCEPTABLE",
              "407 PROXY_AUTHENTICATION_REQUIRED",
              "408 REQUEST_TIMEOUT",
              "409 CONFLICT",
              "410 GONE",
              "411 LENGTH_REQUIRED",
              "412 PRECONDITION_FAILED",
              "413 PAYLOAD_TOO_LARGE",
              "413 REQUEST_ENTITY_TOO_LARGE",
              "414 URI_TOO_LONG",
              "414 REQUEST_URI_TOO_LONG",
              "415 UNSUPPORTED_MEDIA_TYPE",
              "416 REQUESTED_RANGE_NOT_SATISFIABLE",
              "417 EXPECTATION_FAILED",
              "418 I_AM_A_TEAPOT",
              "419 INSUFFICIENT_SPACE_ON_RESOURCE",
              "420 METHOD_FAILURE",
              "421 DESTINATION_LOCKED",
              "422 UNPROCESSABLE_ENTITY",
              "423 LOCKED",
              "424 FAILED_DEPENDENCY",
              "425 TOO_EARLY",
              "426 UPGRADE_REQUIRED",
              "428 PRECONDITION_REQUIRED",
              "429 TOO_MANY_REQUESTS",
              "431 REQUEST_HEADER_FIELDS_TOO_LARGE",
              "451 UNAVAILABLE_FOR_LEGAL_REASONS",
              "500 INTERNAL_SERVER_ERROR",
              "501 NOT_IMPLEMENTED",
              "502 BAD_GATEWAY",
              "503 SERVICE_UNAVAILABLE",
              "504 GATEWAY_TIMEOUT",
              "505 HTTP_VERSION_NOT_SUPPORTED",
              "506 VARIANT_ALSO_NEGOTIATES",
              "507 INSUFFICIENT_STORAGE",
              "508 LOOP_DETECTED",
              "509 BANDWIDTH_LIMIT_EXCEEDED",
              "510 NOT_EXTENDED",
              "511 NETWORK_AUTHENTICATION_REQUIRED"
            ]
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "string"
          },
          "timestamp": {
            "type": "string"
          }
        }
      },
...
      "DeviceRequest": {
        "required": [
          "description",
          "deviceType",
          "manufacturer",
          "name"
        ],
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "installationDate": {
            "type": "string"
          },
          "deviceType": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          }
        }
      },
      "ApiResponseDeviceResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "data": {
            "$ref": "#/components/schemas/DeviceResponse"
          }
        }
      },
      "DeviceResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdAt": {
            "type": "string"
          },
          "createdBy": {
            "type": "string"
          },
          "lastUpdatedAt": {
            "type": "string"
          },
          "lastUpdatedBy": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "installationDate": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "deviceType": {
            "type": "string"
          },
          "manufacturer": {
            "type": "string"
          },
          "active": {
            "type": "boolean"
          }
        }
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "data": {
            "type": "object"
          }
        }
      },
...
      "ApiResponseDeviceActivateResponse": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string"
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "errors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "data": {
            "$ref": "#/components/schemas/DeviceActivateResponse"
          }
        }
      },
      "DeviceActivateResponse": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer Auth": {
        "type": "http",
        "description": "JWT bearer token security scheme",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}